+++ src/RcppExports.cpp
@@ -11,6 +11,27 @@ Rcpp::Rostream<true>&  Rcpp::Rcout = Rcpp::Rcpp_cout_get();
 Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
 #endif
 
+// dann_set_threads_C
+int dann_set_threads_C(int n);
+RcppExport SEXP _dann_dann_set_threads_C(SEXP nSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< int >::type n(nSEXP);
+    rcpp_result_gen = Rcpp::wrap(dann_set_threads_C(n));
+    return rcpp_result_gen;
+END_RCPP
+}
+// dann_get_threads_C
+int dann_get_threads_C();
+RcppExport SEXP _dann_dann_get_threads_C() {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    rcpp_result_gen = Rcpp::wrap(dann_get_threads_C());
+    return rcpp_result_gen;
+END_RCPP
+}
 // dann_predict_all_C
 Rcpp::List dann_predict_all_C(const arma::mat& xTrain, const arma::vec& yTrain, const arma::mat& xTest, int k, int neighborhood_size, double epsilon, const arma::vec& y_class_precedence, const arma::vec& unique_classes, bool probability);
 RcppExport SEXP _dann_dann_predict_all_C(SEXP xTrainSEXP, SEXP yTrainSEXP, SEXP xTestSEXP, SEXP kSEXP, SEXP neighborhood_sizeSEXP, SEXP epsilonSEXP, SEXP y_class_precedenceSEXP, SEXP unique_classesSEXP, SEXP probabilitySEXP) {
@@ -32,6 +53,8 @@ END_RCPP
 }
 
 static const R_CallMethodDef CallEntries[] = {
+    {"_dann_dann_set_threads_C", (DL_FUNC) &_dann_dann_set_threads_C, 1},
+    {"_dann_dann_get_threads_C", (DL_FUNC) &_dann_dann_get_threads_C, 0},
     {"_dann_dann_predict_all_C", (DL_FUNC) &_dann_dann_predict_all_C, 9},
     {NULL, NULL, 0}
 };